+++ src/modules/Accounts/index.js
@@ -15,8 +15,8 @@ import IncognitoSvg from 'assets/svgs/IncognitoSvg';
 import CopyToClipboard from 'components/shared/copyToClipboard';
 import { settingsUpdated } from 'modules/Settings/actions';
 
+import useTransactionList from 'modules/Transactions/hooks/useTransactionList';
 import getStyles from './styles';
-
 import ApplicationSwitcher from '../BlockchainApplication/components/ApplicationSwitcher';
 import { useAccountTokens } from './hooks/useAccounts/useAccountTokens';
 import { useCurrentAccount } from './hooks/useAccounts/useCurrentAccount';
