+++ src/modules/Transactions/components/TransactionDetails/row.js
