+++ src/modules/Transactions/components/TransactionDetails/voteList/index.js
