+++ src/modules/Transactions/components/TransactionDetails/voteList/styles.js
